[quant][graphmode] Fold prepacked weight into module
Summary:
Attach a module to the parent class that contains the packed weight and bias,
this is to support serialization of the quantized model
since the packed weight and bias is not serializable and
we need to overwrite the `__getstate__` and `__setstate__`
function to be able to serialize them

Replace the call to GetAttr[name="weight"] and GetAttr[name="bias"] with
a unpacked weight from the wrapper module.

Next PR will change patterns in QuantFusion.

Test Plan:
python test/test_jit.py

Reviewers:
pt1quant

Subscribers:

Tasks:

Tags:

ghstack-source-id: c556a6b4f0cb21e757fe06d94a8aefdf3e507ebf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26579